+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `table` output mode
+- Filter repositories by group
 
 ## [0.4.0] - 2022-07-29